rename/reorg some files and add plain console option
