+++ src/shared/logging/config.ts
@@ -1,12 +1,13 @@
 import {LOGGLY_URL, LOG_LEVEL} from 'env';
 import {logger} from 'react-native-logs';
 
-import {consoleTransport} from './transports/console-transport';
-import {logglyTransport} from './transports/loggly-transport';
+import {colourConsoleTransport} from './transports/colour-console';
+import {logglyTransport} from './transports/loggly';
+// import {consoleTransport} from './transports/console';
 
 const config = {
   transport: (msg: string | object | Function, level: {severity: number; text: string}, options: any) => {
-    consoleTransport(msg, level, options);
+    colourConsoleTransport(msg, level, options);
     if (LOGGLY_URL) {
       logglyTransport(msg, level, options);
     }
